pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-companions"
version = "2.0.0"
description = "AI Social Simulation Platform with Sentient Autonomous Bots"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "AI Companions Team"}
]
keywords = ["ai", "social", "simulation", "autonomous", "bots", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "redis>=5.0.1",
    "aiohttp>=3.9.1",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "tenacity>=8.2.3",
    "PyYAML>=6.0.1",
    "watchdog>=3.0.0",
    "aiofiles>=23.2.1",
    "Pillow>=10.2.0",
    "prometheus-client>=0.19.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]
celery = [
    "celery>=5.3.6",
]

[project.scripts]
ai-companions = "ai_companions.api.main:main"

[project.urls]
Homepage = "https://github.com/VaibhavJeet/hive"
Documentation = "https://github.com/VaibhavJeet/hive/tree/main/docs"
Repository = "https://github.com/VaibhavJeet/hive"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_companions*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", "__pycache__"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
